Update up/down scripts in arr to the new more robust compose handling.

+++ main/arr/scripts/down-metube.sh
@@ -1,5 +1,20 @@
 #!/bin/bash
 
-# MeTube
-cd /home/arr/metube
+STACK_USER="arr"
+STACK_NAME="metube"
+
+STACK_DIR="/home/$STACK_USER/$STACK_NAME"
+COMPOSE_FILE="$SLSKD_DIR/compose.yaml"
+
+if [[ ! -d "$STACK_DIR" ]]; then
+    echo "[INFO] Directory '$STACK_DIR' does not exist. Skipping $STACK_NAME shutdown."
+    exit 0
+fi
+
+if [[ ! -f "$COMPOSE_FILE" ]]; then
+    echo "[INFO] Compose file '$COMPOSE_FILE' does not exist. Skipping $STACK_NAME shutdown."
+    exit 0
+fi
+
+cd "$STACK_DIR"
 docker compose --env-file ../.env --env-file .env down
